ending?

+++ libftprintf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libftprintf.h                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
+/*   By: jtrancos <jtrancos@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/08 13:19:20 by jtrancos          #+#    #+#             */
-/*   Updated: 2020/10/01 13:07:09 by marvin           ###   ########.fr       */
+/*   Updated: 2020/10/06 12:09:56 by jtrancos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ void	ft_putnbr(int n, t_flags *flags);
 void	ft_print_number(int *n, int num_len, t_flags *flags);
 void	ft_print_x_lower(va_list args, t_flags *flags);
 void	ft_print_x_upper(va_list args, t_flags *flags);
-void	put_hex(unsigned int n, char lettercase, t_flags *flags);
+void	put_hex(size_t n, char lettercase, t_flags *flags);
 int		find_hex_len(unsigned int *n, t_flags *flags);
 void	ft_print_hexnum_lower(unsigned int *n, int hex_len, t_flags *flags);
 void	ft_print_hexnum_upper(unsigned int *n, int hex_len, t_flags *flags);
